cleanup dead code
