+++ packages/plugin-scope-auth/src/steps.ts
@@ -8,8 +8,6 @@ import {
   TypeGrantScopes,
 } from './types';
 
-// TODO generate useful error messages
-
 export function createTypeAuthScopesStep<Types extends SchemaTypes>(
   authScopes: TypeAuthScopes<Types, unknown>,
   type: string,
